+++ pyzx/graph/base.py
@@ -789,13 +789,16 @@ class BaseGraph(Generic[VT, ET], metaclass=DocstringMeta):
                 raise TypeError("Unknown output effect " + s)
         self.set_outputs(tuple(new_outputs))
 
-    def to_tensor(self, preserve_scalar: bool = True, strategy: str = 'naive') -> np.ndarray:
+    def to_tensor(self, preserve_scalar: bool = True, strategy: str = 'naive',
+                  max_memory: Optional[int] = None) -> np.ndarray:
         """Returns a representation of the graph as a tensor using :func:`~pyzx.tensor.tensorfy`"""
-        return tensorfy(self, preserve_scalar, strategy)
+        return tensorfy(self, preserve_scalar, strategy, max_memory=max_memory)
 
-    def to_matrix(self, preserve_scalar: bool = True, strategy: str = 'naive') -> np.ndarray:
+    def to_matrix(self, preserve_scalar: bool = True, strategy: str = 'naive',
+                  max_memory: Optional[int] = None) -> np.ndarray:
         """Returns a representation of the graph as a matrix using :func:`~pyzx.tensor.tensorfy`"""
-        return tensor_to_matrix(tensorfy(self, preserve_scalar, strategy), self.num_inputs(), self.num_outputs())
+        return tensor_to_matrix(tensorfy(self, preserve_scalar, strategy, max_memory=max_memory),
+                                self.num_inputs(), self.num_outputs())
 
     def to_dict(self, include_scalar: bool = True) -> dict[str, Any]:
         """Returns a dictionary representation of the graph, which can then be converted into json."""